+++ src/react/src/__tests__/utils.tests.js
@@ -170,6 +170,7 @@ it('creates a querystring from a set of filter selection', () => {
         contributorTypes: [],
         countries: [],
         sectors: [],
+        statuses: []
     };
 
     const expectedEmptySelectionQSMatch = '';
@@ -266,6 +267,7 @@ it('creates a set of filters from a querystring', () => {
         contributorTypes: [],
         countries: [],
         sectors: [],
+        statuses: [],
         lists: [],
         parentCompany: [],
         facilityType: [],
@@ -300,6 +302,7 @@ it('creates a set of filters from a querystring', () => {
         contributorTypes: [],
         countries: [],
         sectors: [],
+        statuses: [],
         lists: [],
         parentCompany: [],
         facilityType: [],
@@ -325,6 +328,7 @@ it('creates a set of filters from a querystring', () => {
         contributorTypes: [],
         countries: [],
         sectors: [],
+        statuses: [],
         lists: [],
         parentCompany: [],
         facilityType: [],
@@ -355,6 +359,7 @@ it('creates a set of filters from a querystring', () => {
         contributorTypes: [],
         countries: [],
         sectors: [],
+        statuses: [],
         lists: [
             {
                 value: 2,
@@ -394,6 +399,7 @@ it('creates a set of filters from a querystring', () => {
         ],
         countries: [],
         sectors: [],
+        statuses: [],
         lists: [],
         parentCompany: [],
         facilityType: [],
@@ -428,6 +434,7 @@ it('creates a set of filters from a querystring', () => {
             },
         ],
         sectors: [],
+        statuses: [],
         lists: [],
         parentCompany: [],
         facilityType: [],
@@ -458,6 +465,7 @@ it('creates a set of filters from a querystring', () => {
         ],
         countries: [],
         sectors: [],
+        statuses: [],
         lists: [],
         parentCompany: [],
         facilityType: [],
@@ -483,6 +491,7 @@ it('creates a set of filters from a querystring', () => {
         contributorTypes: [],
         countries: [],
         sectors: [],
+        statuses: [],
         lists: [],
         parentCompany: [{
             value: 1,
@@ -515,6 +524,7 @@ it('creates a set of filters from a querystring', () => {
         contributorTypes: [],
         countries: [],
         sectors: [],
+        statuses: [],
         lists: [],
         parentCompany: [],
         facilityType: [{
@@ -547,6 +557,7 @@ it('creates a set of filters from a querystring', () => {
         contributorTypes: [],
         countries: [],
         sectors: [],
+        statuses: [],
         lists: [],
         parentCompany: [],
         facilityType: [],
@@ -579,6 +590,7 @@ it('creates a set of filters from a querystring', () => {
         contributorTypes: [],
         countries: [],
         sectors: [],
+        statuses: [],
         lists: [],
         parentCompany: [],
         facilityType: [],
@@ -611,6 +623,7 @@ it('creates a set of filters from a querystring', () => {
         contributorTypes: [],
         countries: [],
         sectors: [],
+        statuses: [],
         lists: [],
         parentCompany: [],
         facilityType: [],
@@ -643,6 +656,7 @@ it('creates a set of filters from a querystring', () => {
         contributorTypes: [],
         countries: [],
         sectors: [],
+        statuses: [],
         lists: [],
         parentCompany: [],
         facilityType: [],
@@ -660,7 +674,50 @@ it('creates a set of filters from a querystring', () => {
     expect(
         createFiltersFromQueryString(nativeLanguageNameString),
     ).toEqual(expectedNativeLanguageNameMatch);
-    });
+
+    const claimStatusesString = '?statuses=PENDING&statuses=APPROVED&statuses=DENIED&statuses=REVOKED'
+    const expectedClaimStatusesMatch = {
+        facilityFreeTextQuery: '',
+        contributors: [],
+        contributorTypes: [],
+        countries: [],
+        sectors: [],
+        statuses: [
+            {
+                "label": "PENDING",
+                "value": "PENDING",
+            },
+            {
+                "label": "APPROVED",
+                "value": "APPROVED",
+            },
+            {
+                "label": "DENIED",
+                "value": "DENIED",
+            },
+            {
+                "label": "REVOKED",
+                "value": "REVOKED",
+            },
+        ],
+        lists: [],
+        parentCompany: [],
+        facilityType: [],
+        processingType: [],
+        productType: [],
+        numberOfWorkers: [],
+        nativeLanguageName: '',
+        combineContributors: '',
+        boundary: null,
+        sortAlgorithm: {
+            value: 'name_asc', label: 'A to Z',
+          }
+    };
+
+    expect(
+        createFiltersFromQueryString(claimStatusesString),
+    ).toEqual(expectedClaimStatusesMatch);
+});
 
 it('creates a facility detail link', () => {
     const expectedMatch = '/facilities/hello';