+++ src/react/src/components/ClaimedFacilitiesDetails.jsx
@@ -461,7 +461,6 @@ function ClaimedFacilitiesDetails({
                     selectOptions={sectorOptions || []}
                     selectPlaceholder="Select..."
                 />
-                <InputSection label="Product Types" />
                 <InputSection
                     label="Phone Number"
                     value={data.facility_phone_number}