+++ src/react/src/util/constants.jsx
@@ -71,14 +71,18 @@ export const contributorTypeOptions = Object.freeze([
     OTHER,
 ]);
 
-// These choices must be kept in sync with the identical list
-// kept in the Django API's models.py file
 export const facilityClaimStatusChoicesEnum = Object.freeze({
     PENDING: 'PENDING',
     APPROVED: 'APPROVED',
     DENIED: 'DENIED',
     REVOKED: 'REVOKED',
 });
+export const facilityClaimStatusChoices = [
+    { value: facilityClaimStatusChoicesEnum.PENDING, label: 'PENDING' },
+    { value: facilityClaimStatusChoicesEnum.APPROVED, label: 'APPROVED' },
+    { value: facilityClaimStatusChoicesEnum.DENIED, label: 'DENIED' },
+    { value: facilityClaimStatusChoicesEnum.REVOKED, label: 'REVOKED' },
+];
 
 export const facilityListStatusChoicesEnum = Object.freeze({
     PENDING: 'PENDING',