+++ src/react/src/util/util.js
@@ -11,6 +11,7 @@ import omitBy from 'lodash/omitBy';
 import isEmpty from 'lodash/isEmpty';
 import isNumber from 'lodash/isNumber';
 import isNil from 'lodash/isNil';
+import intersection from 'lodash/intersection';
 import values from 'lodash/values';
 import flow from 'lodash/flow';
 import noop from 'lodash/noop';
@@ -28,6 +29,9 @@ import every from 'lodash/every';
 import uniqWith from 'lodash/uniqWith';
 import filter from 'lodash/filter';
 import includes from 'lodash/includes';
+import join from 'lodash/join';
+import map from 'lodash/map';
+import uniq from 'lodash/uniq';
 import { isURL, isInt } from 'validator';
 import { featureCollection, bbox } from '@turf/turf';
 import hash from 'object-hash';
@@ -49,6 +53,7 @@ import {
     DEFAULT_ROWS_PER_PAGE,
     ENTER_KEY,
     facilityListStatusChoicesEnum,
+    facilityClaimStatusChoices,
     facilityListItemStatusChoicesEnum,
     facilityListItemErrorStatuses,
     facilityListSummaryStatusMessages,
@@ -172,6 +177,7 @@ export const makeGetFacilitiesTypeProcessingTypeURL = () =>
     '/api/facility-processing-types/';
 export const makeGetNumberOfWorkersURL = () => '/api/workers-ranges/';
 export const makeGetNativeLanguageName = () => '/api/native_language_name/';
+export const makeGetClaimStatusesURL = () => '/api/claim-statuses/';
 
 export const makeGetFacilitiesURL = () => '/api/facilities/';
 export const makeGetFacilityByOSIdURL = (
@@ -204,7 +210,9 @@ export const makeMergeTwoFacilitiesAPIURL = (targetOSID, toMergeOSID) =>
 export const makeGetFacilitiesCountURL = () => '/api/facilities/count/';
 
 export const makeGetAPIFeatureFlagsURL = () => '/api-feature-flags/';
-export const makeGetFacilityClaimsURL = () => '/api/facility-claims/';
+// TODO: handle &sort_by=contributors_desc at the end of a query
+export const makeGetFacilityClaimsURLWithQueryString = qs =>
+    `/api/facility-claims/?${qs}`;
 export const makeGetFacilityClaimByClaimIDURL = claimID =>
     `/api/facility-claims/${claimID}/`;
 export const makeMessageFacilityClaimantByClaimIDURL = claimID =>
@@ -257,6 +265,7 @@ export const createQueryStringFromSearchFilters = (
         combineContributors = '',
         boundary = {},
         sortAlgorithm = {},
+        claimStatuses = [],
     },
     withEmbed,
     detail,
@@ -269,6 +278,7 @@ export const createQueryStringFromSearchFilters = (
             contributorTypes,
         ),
         countries: createCompactSortedQuerystringInputObject(countries),
+        statuses: createCompactSortedQuerystringInputObject(claimStatuses),
         sectors: createCompactSortedQuerystringInputObject(sectors),
         parent_company: createCompactSortedQuerystringInputObject(
             parentCompany,
@@ -333,6 +343,7 @@ export const createFiltersFromQueryString = qs => {
         contributor_types: contributorTypes = [],
         countries = [],
         sectors = [],
+        statuses = [],
         parent_company: parentCompany = [],
         facility_type: facilityType = [],
         processing_type: processingType = [],
@@ -350,6 +361,7 @@ export const createFiltersFromQueryString = qs => {
         lists: createSelectOptionsFromParams(lists),
         contributorTypes: createSelectOptionsFromParams(contributorTypes),
         countries: createSelectOptionsFromParams(countries),
+        statuses: createSelectOptionsFromParams(statuses),
         sectors: createSelectOptionsFromParams(sectors),
         parentCompany: createSelectOptionsFromParams(parentCompany),
         facilityType: createSelectOptionsFromParams(facilityType),
@@ -419,6 +431,8 @@ export const getTokenFromQueryString = qs => {
     return isArray(token) ? head(token) : token;
 };
 
+const parseFilterQueryString = qs => (startsWith(qs, '?') ? qs.slice(1) : qs);
+
 export const dashboardListParamsDefaults = Object.freeze({
     contributor: null,
     matchResponsibility: matchResponsibilityEnum.MODERATOR,
@@ -426,13 +440,11 @@ export const dashboardListParamsDefaults = Object.freeze({
 });
 
 export const getDashboardListParamsFromQueryString = qs => {
-    const qsToParse = startsWith(qs, '?') ? qs.slice(1) : qs;
-
     const {
         contributor,
         matchResponsibility = dashboardListParamsDefaults.matchResponsibility,
         status = dashboardListParamsDefaults.status,
-    } = querystring.parse(qsToParse);
+    } = querystring.parse(parseFilterQueryString(qs));
 
     return Object.freeze({
         contributor: getNumberFromParsedQueryStringParamOrUseDefault(
@@ -441,6 +453,29 @@ export const getDashboardListParamsFromQueryString = qs => {
         ),
         matchResponsibility,
         status,
+    });
+};
+
+export const dashboardClaimsListParamsDefaults = Object.freeze({
+    countries: [],
+    claimStatuses: facilityClaimStatusChoices[0].value,
+});
+
+export const getDashboardClaimsListParamsFromQueryString = qs => {
+    const {
+        countries,
+        statuses = dashboardClaimsListParamsDefaults.claimStatuses,
+    } = querystring.parse(parseFilterQueryString(qs));
+
+    const statusesArray = Array.isArray(statuses) ? statuses : [statuses];
+    const countriesArray = Array.isArray(countries) ? countries : [countries];
+
+    return Object.freeze({
+        countries: uniq(compact(countriesArray)),
+        statuses: intersection(
+            uniq(compact(statusesArray)),
+            map(facilityClaimStatusChoices, 'value'),
+        ),
     });
 };
 
@@ -675,6 +710,25 @@ export const makeDashboardContributorListLink = ({
     return `/dashboard/lists/${
         params.length > 0 ? `?${params.join('&')}` : ''
     }`;
+};
+
+export const makeDashboardClaimListLink = ({ statuses, countries }) => {
+    const createClaimFilterParams = (key, claimParamValues) =>
+        claimParamValues && claimParamValues.length > 0
+            ? join(
+                  map(claimParamValues, value => `${key}=${value}`),
+                  '&',
+              )
+            : '';
+
+    const statusParams = createClaimFilterParams('statuses', statuses);
+    const countryParams = createClaimFilterParams('countries', countries);
+
+    const params = [statusParams, countryParams]
+        .filter(param => param)
+        .join('&');
+
+    return params ? `/dashboard/claims/?${params}` : '/dashboard/claims';
 };
 
 export const splitContributorsIntoPublicAndNonPublic = contributors =>